feat: seo hot tags
