+++ apps/web/app/hot-tags/components/index.tsx
@@ -1,5 +1,7 @@
 import Rank from './_Rank';
+import Content from './_Content';
 
 export const HotTags = {
   Rank,
+  Content,
 };
